ref: se refactoriza 'MainBottomTabNavigation'
