+++ src/hooks/navigation/index.ts
@@ -1,0 +1,1 @@
+export * from './useMainBottomTabNavigation';
